move renameModel to core
