+++ packages/http-client-java/generator/http-client-generator-core/src/main/java/com/microsoft/typespec/http/client/generator/core/extension/plugin/JavaSettings.java
@@ -3,6 +3,7 @@
 
 package com.microsoft.typespec.http.client.generator.core.extension.plugin;
 
+import com.azure.core.util.CoreUtils;
 import com.azure.json.JsonProviders;
 import com.azure.json.JsonReader;
 import com.azure.json.JsonToken;
@@ -22,6 +23,7 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.Set;
 import java.util.function.Consumer;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import org.slf4j.Logger;
 
@@ -386,6 +388,23 @@ public class JavaSettings {
 
         // Whether to use object for unknown.
         this.useObjectForUnknown = getBooleanValue(host, "use-object-for-unknown", false);
+
+        // Option to rename models (ObjectSchema, ChoiceSchema, SealedChoiceSchema).
+        loadStringSetting("rename-model", s -> {
+            if (!CoreUtils.isNullOrEmpty(s)) {
+                String[] renamePairs = s.split(Pattern.quote(","));
+                for (String pair : renamePairs) {
+                    String[] fromAndTo = pair.split(Pattern.quote(":"));
+                    if (fromAndTo.length == 2) {
+                        String from = fromAndTo[0];
+                        String to = fromAndTo[1];
+                        if (!CoreUtils.isNullOrEmpty(from) && !CoreUtils.isNullOrEmpty(to)) {
+                            this.renameModel.put(from, to);
+                        }
+                    }
+                }
+            }
+        });
     }
 
     private void updateFlavorFactories() {
@@ -1503,6 +1522,12 @@ public class JavaSettings {
 
     public boolean isUseObjectForUnknown() {
         return useObjectForUnknown;
+    }
+
+    private final Map<String, String> renameModel = new HashMap<>();
+
+    public Map<String, String> getJavaNamesForRenameModel() {
+        return renameModel;
     }
 
     private static final String DEFAULT_CODE_GENERATION_HEADER
