+++ packages/http-client-java/generator/http-client-generator-core/src/main/java/com/microsoft/typespec/http/client/generator/core/preprocessor/tranformer/Transformer.java
@@ -67,6 +67,10 @@ public class Transformer {
         } else {
             transformOperationGroups(codeModel.getOperationGroups(), codeModel);
         }
+
+        // process "rename-model"
+        codeModel = new SchemaRenamer(JavaSettings.getInstance().getJavaNamesForRenameModel()).process(codeModel);
+
         return codeModel;
     }
