+++ packages/http-client-java/generator/http-client-generator-mgmt/src/main/java/com/microsoft/typespec/http/client/generator/mgmt/util/FluentJavaSettings.java
@@ -66,8 +66,6 @@ public class FluentJavaSettings {
      */
     private final Map<String, String> namingOverride = new HashMap<>();
 
-    private final Map<String, String> renameModel = new HashMap<>();
-
     private final Set<String> javaNamesForPropertyIncludeAlways = new HashSet<>();
 
     private final Map<String, String> renameOperationGroup = new HashMap<>();
@@ -119,10 +117,6 @@ public class FluentJavaSettings {
 
     public Map<String, String> getNamingOverride() {
         return namingOverride;
-    }
-
-    public Map<String, String> getJavaNamesForRenameModel() {
-        return renameModel;
     }
 
     public Set<String> getJavaNamesForRemoveModel() {
@@ -177,22 +171,6 @@ public class FluentJavaSettings {
         loadStringSetting("add-inner", s -> splitStringToSet(s, javaNamesForAddInner));
 
         loadStringSetting("remove-inner", s -> splitStringToSet(s, javaNamesForRemoveInner));
-
-        loadStringSetting("rename-model", s -> {
-            if (!CoreUtils.isNullOrEmpty(s)) {
-                String[] renamePairs = s.split(Pattern.quote(","));
-                for (String pair : renamePairs) {
-                    String[] fromAndTo = pair.split(Pattern.quote(":"));
-                    if (fromAndTo.length == 2) {
-                        String from = fromAndTo[0];
-                        String to = fromAndTo[1];
-                        if (!CoreUtils.isNullOrEmpty(from) && !CoreUtils.isNullOrEmpty(to)) {
-                            renameModel.put(from, to);
-                        }
-                    }
-                }
-            }
-        });
 
         loadStringSetting("remove-model", s -> splitStringToSet(s, javaNamesForRemoveModel));
